+++ include/qalgorithms_read_file.h
@@ -95,8 +95,8 @@ namespace qAlgorithms
             std::vector<double> *const spectrum_RT,
             size_t index);
 
-        std::vector<size_t> get_spectra_index(const std::vector<unsigned int> *indices,
-                                              const std::vector<pugi::xml_node> *spectra_nodes_ex);
+        std::vector<unsigned int> get_spectra_index(const std::vector<unsigned int> *indices,
+                                                    const std::vector<pugi::xml_node> *spectra_nodes_ex);
         std::vector<int> get_spectra_level(const std::vector<unsigned int> *indices,
                                            const std::vector<pugi::xml_node> *spectra_nodes_ex);
         std::vector<bool> get_spectra_mode(const std::vector<unsigned int> *indices,
